requirements.txt
praw
python-dotenv
anthropic
prawcore
requests
beautifulsoup4
Pillow
